undo complex change

+++ tests/ops/functions/test_dot.py
@@ -232,13 +232,13 @@ class TestDotPauliSentence:
         """Test the dot function with the tensorflow interface."""
         import tensorflow as tf
 
-        c = tf.constant([1.0, 2.0, 3.0], dtype=tf.complex64)
+        c = tf.constant([1.0, 2.0, 3.0])
         o = [qml.PauliX(0), qml.PauliY(1), qml.PauliZ(2)]
         ps = qml.dot(c, o, pauli=True)
         op_sum = Sum(
             qml.PauliX(0),
-            SProd(tf.constant(2.0, dtype=tf.complex64), qml.PauliY(1)),
-            SProd(tf.constant(3.0, dtype=tf.complex64), qml.PauliZ(2)),
+            SProd(tf.constant(2.0), qml.PauliY(1)),
+            SProd(tf.constant(3.0), qml.PauliZ(2)),
         )
         ps_2 = qml.pauli.pauli_sentence(op_sum)
         assert ps == ps_2
